+++ libft/ft_strmapi.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   ft_strmapi.c                                       :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: rzhdanov <rzhdanov@student.42barcelona.co  +#+  +:+       +#+        */
+/*   By: rzhdanov <rzhdanov@student.42barcelona.    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/02/09 01:34:29 by rzhdanov          #+#    #+#             */
-/*   Updated: 2023/03/03 19:45:09 by rzhdanov         ###   ########.fr       */
+/*   Updated: 2024/08/12 05:49:07 by rzhdanov         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,9 +14,9 @@
 
 char	*ft_strmapi(char const *s, char (*f)(unsigned int, char))
 {
-	unsigned int	i;
-	unsigned int	limit;
-	char			*result;
+	size_t	i;
+	size_t	limit;
+	char	*result;
 
 	i = 0;
 	limit = ft_strlen(s);
